requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
streamlit==1.28.2
requests==2.31.0
pandas==2.1.3
scikit-learn==1.3.2
numpy==1.26.2
PyPDF2==3.0.1
python-docx==1.1.0
